Revert "lockhammer/prefetch64: do not dereference lock pointer"
My analysis that resulted in the previous commit c2db765 was not complete.  The
causes of the additional memory accesses come from elsewhere.  The patch
actually does the wrong thing that I was trying to fix.  The patch causes code
to be geneated that pushes the value of the pointer onto the stack and then a
prefetch from the address of that entry.  The "Q" constraint creates the
base-register memory address for an lvalue and works as intended.

This reverts commit c2db765d1bab8b80642a8c135f763895ae0b3d38.