mysql/cas_event_mutex.h: fix assembly constraints
Compiler warns that transfer and address registers are the same,
but this is unpredictable according to architecture.  Add the &
modifier for earlyclobber to use a different register for the
transfer register than the lock address register.  Also, use Q
constraint for the lock address as a single base register with
no offset.

Change-Id: If9ccafd401a8d68b763d1eb5be1b0e52ad07ada2