lockhammer: add CMN HN-F targeting
On CMN-based systems, SLC and memory latency varies by HN-F device,
the selection of which depends on hashing the physical address.
On large meshes, the latency variability can be significantly large.
Therefore, a lock variable's memory address affects performance.

This patch provides a way to target the HN-F node in which to
allocate the locks so that performance can be reliably reproducible.

These new flags control the HN-F placement.

-f select<#> : algorithm to use for computing HN-F select value
-H value : HN-F select value to use to target
-F value : number of lock search tries

Note cmn-phys-mem.c provides the hash for 64 HN-Fs, which is not
yet documented in the CMN-600 TRM.

Change-Id: I199ca38c5930ae76786cdb57e47c4cb7a18db786