lockhammer: improve test of valid online core
The -o flag lets any CPU number to be specified to run a worker
thread, but the CPU may not be online due to CPU hotplugging.
This patch gets the set of online cores to check if the core is
online instead of assuming the numbering of online cores is
contiguous.

Change-Id: I08851882b4740d8e20d5dfba3e0abcdd7df686f0